Move PageDto to new format
