+++ apps/api/src/models/inventory/inventory.repository.ts
@@ -1,10 +1,10 @@
 import { Injectable } from '@nestjs/common';
 import { InjectModel } from '@nestjs/mongoose';
 import mongoose, { FilterQuery, Model, Types } from 'mongoose';
-import { PageQueryDto } from 'shared-types';
 import { EntityRepository } from '../../database/entity.repository';
 import { OrgValueCalculationStrategy } from '../organizations/schemas/org-settings';
 import { InventoryItem, InventoryItemDocument } from './schemas/inventory-item.schema';
+import PageQueryDto from 'shared-types/dist/page/IPageQueryDto';
 
 @Injectable()
 export class InventoryRepository extends EntityRepository<InventoryItemDocument> {
