+++ apps/api/src/models/products/products.service.ts
@@ -1,6 +1,5 @@
 import { Injectable } from '@nestjs/common';
 import mongoose, { Types } from 'mongoose';
-import { PageQueryDto } from 'shared-types';
 import { RepositoryPaginateResult } from '../../database/entity.repository';
 import { ImagesService } from '../../images/images.service';
 import { InventoryService } from '../inventory/inventory.service';
@@ -8,6 +7,7 @@ import { ProductsRepository } from './products.repository';
 import { ProductDocument } from './schemas/product.schema';
 import CreateProductDto from './dto/create-product-dto';
 import UpdateProductDto from './dto/update-product-dto';
+import PageQueryDto from '../../dto/page-query-dto';
 
 @Injectable()
 export class ProductsService {
