+++ packages/shared-types/src/product/IUpdateProductDto.ts
